+++ src/textual/_types.py
@@ -1,6 +1,12 @@
 from typing import TYPE_CHECKING, Any, Awaitable, Callable, List, Union
 
-from typing_extensions import Protocol
+from typing_extensions import (
+    Literal,
+    Protocol,
+    SupportsIndex,
+    get_args,
+    runtime_checkable,
+)
 
 if TYPE_CHECKING:
     from rich.segment import Segment